test(bevy): 寻路/坐标转换单元测试（24 全过）+ clippy 清理
- pathfinding：直线/绕障/不可达/同格 4 测试
- movement：tile↔world 往返 + 方向 2 测试
- 清理新增文件 clippy 警告（未用导入/括号/cast/dead_code）

+++ Client-Bevy/src/game/weather.rs
@@ -120,8 +120,8 @@ fn fastrand_f32() -> f32 {
     let seed = SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .map(|d| d.subsec_nanos())
-        .unwrap_or(0) as u32;
-    ((seed.wrapping_mul(1664525).wrapping_add(1013904223)) as f32 / u32::MAX as f32)
+        .unwrap_or(0);
+    (seed.wrapping_mul(1664525).wrapping_add(1013904223)) as f32 / u32::MAX as f32
 }
 
 /// 推进雨/雪粒子
